General code cleanup. Improve editing for GNDVNode
Fix possible bug where a UTF8 string could be read outside of intended boundaries for BLOCEntries

+++ BrawlLib/SSBB/ResourceNodes/ProjectPlus/ASLSNode.cs
@@ -131,7 +131,7 @@ namespace BrawlLib.SSBB.ResourceNodes.ProjectPlus
         {
             _buttonFlags = Header->_buttonFlags;
             _name = Parent.WorkingUncompressed.Address.GetUTF8String(
-                ((ASLSNode) Parent).Header->_nameOffset + Header->_nameOffset);
+                (uint)((ASLSNode) Parent).Header->_nameOffset + Header->_nameOffset);
             return false;
         }
     }
